oblimap_par_gcm_to_im: port scan phase to mpi shared arrays
This patch introduces mpi_helper.f90 which defines 2 subroutines:
- decompose: returns the slice n0:n1 of an array on proc p given a global size N amongst M processors
- alloc_shared: returns a shared array a_sm allocated on the Proc 0 and returns the slice
a_ on  the current processor.

Accessing the slice a_sm(n0:n1) is the same as acessing a_(n0:n1).
Concurrent writing to the same location(s) of array could lead to undefine values...

For now, the GCM domain is splitted along the NX dimension but the code is prepared to a 2D
decomposition (in both NX and NY).

NB: some other changes has been done in this patch as
- doing the loop inside the statements of a select/case
- filling the arrays with the contiguous dimension in the inner loop (i or n) and the second dimension
(j or m) as the outer loop.